requirements.txt
aniso8601==9.0.1
APScheduler==3.6.3
blis==0.7.4
cachetools==4.2.2
catalogue==2.0.6
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
cymem==2.0.5
Flask==2.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
murmurhash==1.0.5
numpy==1.21.2
packaging==21.0
pandas==1.3.2
pathy==0.6.0
Pillow==8.3.2
preshed==3.0.5
pydantic==1.8.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
six==1.16.0
smart-open==5.2.1
spacy==3.1.2
spacy-legacy==3.0.8
SQLAlchemy==1.4.23
SQLAlchemy-Utils==0.37.8
srsly==2.4.1
thinc==8.0.9
torch==1.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.62.2
typer==0.3.2
typing-extensions==3.10.0.2
tzlocal==3.0
urllib3==1.26.6
wasabi==0.8.2
Werkzeug==2.0.1
